feat(global_list): add support for global lists
Implementing support for lists across all collections. The
motivation is to prove that it is indeed possible to use the "**"
characters to signify a global read: https://github.com/aep-dev/aep.dev/pull/165/files#r1565202289.

After this PR, GlobalList, List, and Read all map to the appropriate
collections:

```
$ curl 'http://localhost:8081/publisher/foo/book/bar'
{"code":5, "message":"book \"publisher/foo/book/bar\" not found", "details":[]}
$ curl 'http://localhost:8081/publisher/**/book'
{"code":12, "message":"method ListBook not implemented", "details":[]}
$ curl 'http://localhost:8081/**/book'
{"code":12, "message":"method GlobalListBook not implemented", "details":[]}
```

Also fixing a bug with update not using the PUT HTTP method.