fix: get end-to-end CRUD journey, fix grpc
The example server new supports an end-to-end
CRUD flow of creating, reading, updating, and
deleting a resource.

This serves as a great illustrative example of 
a resource-oriented API as well as validating 
the function of the test server.

grpc gateway integration was also broken when
the generated file structure was changed, so that
is now fixed.