feat: multiple nesting, deduped collection names
The AEPs allow deduplicating collection names (for example, 
removing the "book" from "book-editions" in the path). Since
this is a fairly common practice, make it something that aepc
will do for you by default.

Also fixing a few places where multiple nested children was not working,
as well as additional issues that came from a hyphen in the resource singular.

Also started building out an internal/ package - eventually most packages
will move there, as aepc is a first and foremost a command-line interface
and intended to be used as a binary.

A utils library was added to do basic casing, as it is common to 
do so.