feat(service): add list skip / CEL filters, test db
To provide an illustrative example of CEL's usage as a filter language,
implement it, along with filtering and sending data along to a functional database
(in this case, sqlite).

A small example cel2sql parser was written to illustrate the ability to use
CEL as a query language. However, this will be replaced by some cel -> sql parser,
hopefully contributed as a more general library and dependency.

Example:

```
curl $'http://localhost:8081/publishers?filter=description.startsWith(\'tomorrow\')'
{"results":[{"description":"tomorrow they will write again","path":"publishers/foo"}],"nextPageToken":""}
```

Fixes #59