feat(151): add long-running-operation example (#66)
Adding an example of a long-running-operation (archiveBook), to illustrate how LROs would work from a user interface standpoint.

This change currently implements the google.aip.Operation service to return the operation resource - in the future this will be replaced by an operation service that will be generated by aep-lib-go/aepc.

This change also upstreams the proto generation functionality, as well as the resource model for aepc, up to aep-lib-go. This simplifies end-to-end testing of changes to the resource model, as well as it's maintenance.